Optimize string table writing, for Bug #990: Transactions spreadsheet download is very slow

+++ YoFi.Tests/NewSpreadsheetsTest.cs
@@ -34,7 +34,7 @@ namespace YoFi.Tests
             }
         }
 
-        void WhenWritingToNewSpreadsheet<T>(Stream stream,IEnumerable<T> items, string name) where T: class
+        void WhenWritingToNewSpreadsheet<T>(Stream stream,IEnumerable<T> items, string name,bool writetodisk = true) where T: class
         {
             using (var writer = new OpenXmlSpreadsheetWriter())
             {
@@ -43,12 +43,16 @@ namespace YoFi.Tests
             }
 
             stream.Seek(0, SeekOrigin.Begin);
-            var filename = $"Test-{name}.xlsx";
-            File.Delete(filename);
-            using (var outstream = File.OpenWrite(filename))
+
+            if (writetodisk)
             {
-                Console.WriteLine($"Writing {outstream.Name}...");
-                stream.CopyTo(outstream);
+                var filename = $"Test-{name}.xlsx";
+                File.Delete(filename);
+                using (var outstream = File.OpenWrite(filename))
+                {
+                    Console.WriteLine($"Writing {outstream.Name}...");
+                    stream.CopyTo(outstream);
+                }
             }
         }
 
@@ -78,14 +82,14 @@ namespace YoFi.Tests
             }
         }
 
-        public void WriteThenReadBack<T>(string name, IEnumerable<T> items, bool newreader = false) where T : class, new()
+        public void WriteThenReadBack<T>(string name, IEnumerable<T> items, bool newreader = false, bool writetodisk = true) where T : class, new()
         {
             // Given: Some items
 
             // When: Writing it to a spreadsheet using the new methods
             using (var stream = new MemoryStream())
             {
-                WhenWritingToNewSpreadsheet(stream, items, name);
+                WhenWritingToNewSpreadsheet(stream, items, name, writetodisk);
 
                 // And: Reading it back to a spreadsheet using the old methods
                 var actual = new List<T>();
@@ -304,7 +308,7 @@ namespace YoFi.Tests
             // When: Writing it to a spreadsheet using the new methods
             // And: Reading it back to a spreadsheet using the old methods
             // Then: The spreadsheet is valid, and contains the expected item
-            WriteThenReadBack<Transaction>("TransactionItems200", Items, newreader:true);
+            WriteThenReadBack<Transaction>("TransactionItems200", Items, newreader:true, writetodisk:false);
         }
 
         [DataRow(true)]
